+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Minor*
 
 * Renamed the Defender to Protector. ([#132](https://github.com/hjwylde/werewolf/issues/132))
+* Automatically delete the game file if the game is over. ([#100](https://github.com/hjwylde/werewolf/issues/100))
 
 *Revisions*